% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hist_dens.R
\name{create_hist_dens}
\alias{create_hist_dens}
\title{Class Constructor Function: create_hist_dens}
\usage{
create_hist_dens(x, type = c("gaussian", "kde", "fp"), ...)
}
\arguments{
\item{x}{A numeric vector representing the data}

\item{type}{Character string specifying the type of density estimation to be
used. Options are "gaussian", "kde", and "fp"}

\item{...}{Additional arguments to be passed to the density estimation
function}
}
\value{
An object of class \code{"hist_dens"} containing the
histogram (\code{h}), density values (\code{d}) and type of density
estimation used (\code{type}).
}
\description{
Function computes a histogram and chooses the density estimation provided by
user (default=Gaussian), assigning the result to class \code{"hist_dens"}
}
\examples{
       create_hist_dens(mtcars$mpg, type="kde", bw=2)
       create_hist_dens(iris$Sepal.Length, type="fp")

}
\seealso{
\code{\link[stats]{density}} for a complete list of arguments that
        can be passed when \code{type = "kde"}
         \code{\link{frequency_polygon}} to create frequency polygon density estimations
          when type specified is \code{type = "fp"}

Other densr: 
\code{\link{frequency_polygon}()},
\code{\link{plot.hist_dens}()}
}
\author{
Rachel McInerney - <\email{rachel.mcinerney.2021@mumail.ie}>
        Conor Thompson - <\email{conor.thompsongargan.2020@mumail.ie}>
        Osi Ukachukwu - <\email{osi.ukachukwu.2018@mumail.ie}>
}
\concept{densr}
