% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densr.R
\docType{package}
\name{.onAttach}
\alias{.onAttach}
\title{densr}
\usage{
.onAttach(lib, pkg)
}
\description{
A package to overlay a histogram with different types of density estimation methods on a numeric data set.
}
\section{Details}{

\describe{
\item{Type: }{Package}
\item{Package: }{densr}
\item{Version: }{0.1.0}
\item{Date: }{2024-01-04}
}
}

\section{Usage}{

Package contains three functions which are to be used in conjunction.
These are \code{link{frequency_polygon}}, for creating a frequency polygon
type density estimation, \code{link{create_hist_dens}}, to create an object
of class \code{"hist_dens"}, \code{link{plot.hist_dens}}, a dedicated plotting function
is used to plot objects of class \code{"hist_dens"}
}

\examples{
NULL

}
\author{
Rachel McInerney - <\email{rachel.mcinerney.2021@mumail.ie}>
        Conor Thompson - <\email{conor.thompsongargan.2020@mumail.ie}>
        Osi Ukachukwu - <\email{osi.ukachukwu.2018@mumail.ie}>
}
\keyword{package}
