% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hist_dens.R
\name{plot.hist_dens}
\alias{plot.hist_dens}
\title{plot.hist_dens Method}
\usage{
\method{plot}{hist_dens}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"hist_dens"}}

\item{...}{Different possible arguments that may be passed to \code{plot},
\code{curve}, or \code{lines}}
}
\value{
A histogram of numeric data with an explicitly stated density option
superimposed on it.
}
\description{
A method defined to plot an object of class \code{"hist_dens"}.
}
\details{
This function overlays a density estimate, based on the
type specified i.e. Gaussian, kde, or fp, to a histogram of the data \code{x}
supplied to \code{\link{create_hist_dens}}.
}
\note{
For this plot method to run efficiently,
the argument \code{x} must be of class \code{"hist_dens"}
}
\examples{
x1 <- create_hist_dens(mtcars$mpg, type="kde", bw=2)
plot(x1)
plot(create_hist_dens(iris$Sepal.Length, type="fp"))

}
\seealso{
\code{\link{create_hist_dens}} to create objects of class
\code{"hist_dens"}

Other densr: 
\code{\link{create_hist_dens}()},
\code{\link{frequency_polygon}()}
}
\author{
Rachel McInerney - <\email{rachel.mcinerney.2021@mumail.ie}>
        Conor Thompson - <\email{conor.thompsongargan.2020@mumail.ie}>
        Osi Ukachukwu - <\email{osi.ukachukwu.2018@mumail.ie}>
}
\concept{densr}
