% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hist_dens.R
\name{plot}
\alias{plot}
\title{Plot create_hist_dens() output}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{An object of class \code{"hist_dens"} generated by
\code{create_hist_dens}.}

\item{...}{Catches unused arguments to \code{plot}, \code{curve}.
\code{dnorm} or \code{lines}.}
}
\value{
Nothing in the Console. A histogram of numeric data with an
explicitly stated density option superimposed on it is produced in the Plots
window.
}
\description{
Plotting method for objects inheriting from the class \code{"hist_dens"}.
Overlays a density estimate based on the type specified in
\code{create_hist_dens} i.e. gaussian, kde, or fp, to a histogram of the
same data supplied to \code{create_hist_dens}.
}
\examples{
dat1 <- create_hist_dens(ChickWeight$weight, type = "kde", kernel = "epanechnikov")
plot(dat1, lwd = 2)
dat2 <- create_hist_dens(USArrests$Assault, type = "fp")
plot(dat2, lty = 4)
}
\seealso{
\code{\link{create_hist_dens}} for how to create objects of class
\code{"hist_dens"}

Other densr: 
\code{\link{create_hist_dens}()},
\code{\link{frequency_polygon}()}
}
\author{
Rachel McInerney - <\email{rachel.mcinerney.2021@mumail.ie}>,
        Conor Thompson - <\email{conor.thompsongargan.2020@mumail.ie}>,
        Osi Ukachukwu - <\email{osi.ukachukwu.2018@mumail.ie}>
}
\concept{densr}
