% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_polygon.R
\name{frequency_polygon}
\alias{frequency_polygon}
\title{Frequency Polygon}
\usage{
frequency_polygon(x, ...)
}
\arguments{
\item{x}{A numeric vector of data values}

\item{...}{Additional arguments to be passed to the \code{hist} function}
}
\value{
An object of class \code{list} containing coordinates for plotting
the frequency polygon. Midpoint of density histogram represented by the
x-coordinate (\code{vx}), and the corresponding density values represented
by the y-coordinate (\code{vy})
}
\description{
Calculates the coordinates to plot the frequency polygon density estimate.
This function adapts code from the notes to create a frequency polygon by
extracting components of a density histogram object and estimating density
at locations excluding midpoints.
}
\examples{
fp <- frequency_polygon(x = mtcars$mpg)

}
\seealso{
\code{\link{hist}}

Other densr: 
\code{\link{create_hist_dens}()},
\code{\link{plot.hist_dens}()}
}
\author{
Rachel McInerney - <\email{rachel.mcinerney.2021@mumail.ie}>
        Conor Thompson - <\email{conor.thompsongargan.2020@mumail.ie}>
        Osi Ukachukwu - <\email{osi.ukachukwu.2018@mumail.ie}>
}
\concept{densr}
