% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hist_dens.R
\name{plot.hist_dens}
\alias{plot.hist_dens}
\title{Histogram density overlay plot method}
\usage{
\method{plot}{hist_dens}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"hist_dens"}}

\item{...}{Additional arguments to be passed to}
}
\description{
Method created to plot an object of class \code{"hist_dens"} using the
 \code{plot.hist_dens}. This method overlays the histogram and density
 estimates based on the specified type, ie. Gaussian, kde, or fp.
}
\note{
The input \code{x} must be of class \code{"hist_dens"}
}
\examples{
plot(create_hist_dens(mtcars$mpg, type="kde", bw=2))
plot(create_hist_dens(iris$Sepal.Length, type="fp"))

}
\seealso{
\code{\link{create_hist_dens}} to create objects of class
\code{"hist_dens"}

Other densr: 
\code{\link{create_hist_dens}()},
\code{\link{frequency_polygon}()}
}
\author{
Rachel McInerney - <\email{rachel.mcinerney.2021@mumail.ie}>
        Conor Thompson - <\email{conor.thompsongargan.2020@mumail.ie}>
        Osi Ukachukwu - <\email{osi.ukachukwu.2018@mumail.ie}>
}
\concept{densr}
